+++ Craftsman/Builders/Tests/IntegrationTests/ConsumerTestBuilder.cs
@@ -30,7 +30,7 @@ public class ConsumerTestBuilder
         return @$"namespace {classPath.ClassNamespace};
 
 using FluentAssertions;
-using NUnit.Framework;
+using Xunit;
 using System.Threading.Tasks;
 using MassTransit;
 using MassTransit.Testing;
@@ -41,6 +41,7 @@ using {consumerClassPath.ClassNamespace};
 using {testUtilClassPath.ClassNamespace};
 using static {testFixtureName};
 
+[Collection(nameof(TestFixture))]
 public class {consumer.ConsumerName}Tests : TestBase
 {{
     {ConsumerTest(consumer)}
@@ -51,7 +52,7 @@ public class {consumer.ConsumerName}Tests : TestBase
     {
         var messageName = FileNames.MessageInterfaceName(consumer.MessageName);
 
-        return $@"[Test]
+        return $@"[Fact]
     public async Task can_consume_{consumer.MessageName}_message()
     {{
         // Arrange