+++ Craftsman/Builders/Tests/IntegrationTests/DeleteCommandTestBuilder.cs
@@ -43,11 +43,12 @@ using {testUtilClassPath.ClassNamespace};
 using {featuresClassPath.ClassNamespace};
 using FluentAssertions;
 using Microsoft.EntityFrameworkCore;
-using NUnit.Framework;
+using Xunit;
 using {exceptionsClassPath.ClassNamespace};
 using System.Threading.Tasks;
 using static {testFixtureName};{foreignEntityUsings}
 
+[Collection(nameof(TestFixture))]
 public class {commandName}Tests : TestBase
 {{
     {GetDeleteTest(commandName, entity, featureName)}{GetDeleteWithoutKeyTest(commandName, entity, featureName)}{softDeleteTest}
@@ -66,7 +67,7 @@ public class {commandName}Tests : TestBase
 
         var fakeParent = IntegrationTestServices.FakeParentTestHelpers(entity, out var fakeParentIdRuleFor);
 
-        return $@"[Test]
+        return $@"[Fact]
     public async Task can_delete_{entity.Name.ToLower()}_from_db()
     {{
         // Arrange
@@ -91,7 +92,7 @@ public class {commandName}Tests : TestBase
 
         return badId == "" ? "" : $@"
 
-    [Test]
+    [Fact]
     public async Task delete_{entity.Name.ToLower()}_throws_notfoundexception_when_record_does_not_exist()
     {{
         // Arrange
@@ -119,7 +120,7 @@ public class {commandName}Tests : TestBase
 
         return $@"
 
-    [Test]
+    [Fact]
     public async Task can_softdelete_{entity.Name.ToLower()}_from_db()
     {{
         // Arrange