+++ Craftsman/Builders/Tests/IntegrationTests/PatchCommandTestBuilder.cs
@@ -58,11 +58,12 @@ using {exceptionClassPath.ClassNamespace};
 using {featuresClassPath.ClassNamespace};
 using FluentAssertions;
 using Microsoft.EntityFrameworkCore;
-using NUnit.Framework;
+using Xunit;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.JsonPatch;
 using static {testFixtureName};{foreignEntityUsings}
 
+[Collection(nameof(TestFixture))]
 public class {commandName}Tests : TestBase
 {{
     {GetAddCommandTest(commandName, entity, featureName, lookupVal, myProp)}
@@ -82,7 +83,7 @@ public class {commandName}Tests : TestBase
 
         var fakeParent = IntegrationTestServices.FakeParentTestHelpers(entity, out var fakeParentIdRuleFor);
 
-        return $@"[Test]
+        return $@"[Fact]
     public async Task can_patch_existing_{entity.Name.ToLower()}_in_db()
     {{
         // Arrange
@@ -112,7 +113,7 @@ public class {commandName}Tests : TestBase
 
         return randomId == "" ? "" : $@"
 
-    [Test]
+    [Fact]
     public async Task passing_null_patchdoc_throws_validationexception()
     {{
         // Arrange
@@ -133,7 +134,7 @@ public class {commandName}Tests : TestBase
         var updateDto = FileNames.GetDtoName(entity.Name, Dto.Update);
 
         return badId == "" ? "" : $@"
-    [Test]
+    [Fact]
     public async Task patch_{entity.Name.ToLower()}_throws_notfound_exception_when_record_does_not_exist()
     {{
         // Arrange